Migrate missing function and imports
