+++ datameta/api/ui/forgot.py
@@ -15,12 +15,27 @@
 from pyramid.httpexceptions import HTTPFound
 from pyramid.view import view_config
 
-from ... import security
+from ... import security, email
+from ...settings import get_setting
 
 import logging
 log = logging.getLogger(__name__)
 
 import re
+
+def send_forgot_token(request, db_token_obj, clear_token):
+    """Sends a password reset token email to the corresponding user"""
+    db = request.dbsession
+
+    email.send(
+        recipients = (db_token_obj.user.fullname, db_token_obj.user.email),
+        subject = get_setting(db, "subject_forgot_token").str_value,
+        template = get_setting(db, "template_forgot_token").str_value,
+        values={
+           'fullname' : db_token_obj.user.fullname,
+           'token_url' : request.route_url('setpass', token = clear_token)
+           }
+        )
 
 @view_config(route_name='forgot_api', renderer='json')
 def v_forgot_api(request):
